Implementation of API, repository, VM, model and View for getting of all pots by greenhouse ID

+++ app/src/main/java/com/example/greehousecontroller/data/model/Pot.java
@@ -3,17 +3,17 @@ package com.example.greehousecontroller.data.model;
 public class Pot {
     private int id;
     private String name;
-    private int currentHumidity;
-    private int minimalHumidity;
+    private float currentHumidity;
+    private float minimalHumidity;
 
-    public Pot(int id, String name, int currentHumidity, int minimalHumidity) {
+    public Pot(int id, String name, float currentHumidity, float minimalHumidity) {
         this.id = id;
         this.name = name;
         this.currentHumidity = currentHumidity;
         this.minimalHumidity = minimalHumidity;
     }
 
-    public Pot(String name, int currentHumidity, int minimalHumidity) {
+    public Pot(String name, float currentHumidity, float minimalHumidity) {
         this.name = name;
         this.currentHumidity = currentHumidity;
         this.minimalHumidity = minimalHumidity;
@@ -27,11 +27,11 @@ public class Pot {
         return name;
     }
 
-    public int getCurrentHumidity() {
+    public float getCurrentHumidity() {
         return currentHumidity;
     }
 
-    public int getMinimalHumidity(){
+    public float getMinimalHumidity(){
         return minimalHumidity;
     }
 }
